+++ src/index.js
@@ -5,13 +5,21 @@ import { applySql, applySqlFile, applySqlFiles, getMigrations, simulate } from '
 import Database from './sim/Database';
 import type {
   BinaryDataType,
+  BinaryWithLengthDataType,
+  BinaryWithoutLengthDataType,
   DataType,
+  DateOnlyDataType,
   DateTimeDataType,
   EnumDataType,
-  IntDataType,
-  OtherDataType,
-  RealDataType,
-  TextDataType,
+  FixedPointDataType,
+  FloatingPointDataType,
+  JSONDataType,
+  NumericDataType,
+  StringDataType,
+  StringOrEnumDataType,
+  StringWithLengthDataType,
+  StringWithoutLengthDataType,
+  TimeOnlyDataType,
 } from './sim/DataType';
 import ForeignKey from './sim/ForeignKey';
 import Index from './sim/Index';
@@ -20,14 +28,21 @@ import Table from './sim/Table';
 
 export type {
   BinaryDataType,
+  BinaryWithLengthDataType,
+  BinaryWithoutLengthDataType,
   DataType,
+  DateOnlyDataType,
   DateTimeDataType,
   EnumDataType,
-  IndexType,
-  IntDataType,
-  OtherDataType,
-  RealDataType,
-  TextDataType,
+  FixedPointDataType,
+  FloatingPointDataType,
+  JSONDataType,
+  NumericDataType,
+  StringDataType,
+  StringOrEnumDataType,
+  StringWithLengthDataType,
+  StringWithoutLengthDataType,
+  TimeOnlyDataType,
 };
 export { applySql, applySqlFile, applySqlFiles, getMigrations, simulate, Column, Database, ForeignKey, Index, Table };
 export type { MigrationInfo } from './sim/core';