+++ src/sim/DataType.js
@@ -4,15 +4,19 @@ import type { Encoding } from './encodings';
 import { getDefaultCollationForCharset, makeEncoding } from './encodings';
 import { parseEnumValues, quote } from './utils';
 
-export type IntDataType = {
+export type IntegerDataType = {
+  category: 'NUMERIC',
+  subCategory: 'INTEGER',
   baseType: 'tinyint' | 'smallint' | 'mediumint' | 'int' | 'bigint',
   length: number,
   unsigned: boolean,
   zeroFill: boolean,
 };
 
-export type RealDataType = {
-  baseType: 'float' | 'double' | 'decimal',
+export type FixedPointDataType = {
+  category: 'NUMERIC',
+  subCategory: 'FIXED_POINT',
+  baseType: 'decimal',
   precision: {
     length: number,
     decimals: number,
@@ -21,43 +25,88 @@ export type RealDataType = {
   zeroFill: boolean,
 };
 
+export type FloatingPointDataType = {
+  category: 'NUMERIC',
+  subCategory: 'FLOATING_POINT',
+  baseType: 'float' | 'double',
+  precision: {
+    length: number,
+    decimals: number,
+  } | null,
+  unsigned: boolean,
+  zeroFill: boolean,
+};
+
+export type NumericDataType = IntegerDataType | FixedPointDataType | FloatingPointDataType;
+
 export type DateTimeDataType = {
-  // NOTE: "DATE" does not belong here! It's an OtherDataType, as it does not
-  // have any parameters!
-  baseType: 'time' | 'timestamp' | 'datetime',
+  category: 'TEMPORAL',
+  baseType: 'timestamp' | 'datetime',
   fsp: number | null,
 };
 
-export type TextDataType = {
-  baseType: 'char' | 'varchar' | 'text' | 'mediumtext' | 'longtext',
-  length: number | null,
+export type DateOnlyDataType = {
+  category: 'TEMPORAL',
+  baseType: 'date' | 'year',
+};
+
+export type TimeOnlyDataType = {
+  category: 'TEMPORAL',
+  baseType: 'time',
+  fsp: number | null,
+};
+
+export type TemporalDataType = DateTimeDataType | DateOnlyDataType | TimeOnlyDataType;
+
+export type StringWithLengthDataType = {
+  category: 'STRING',
+  baseType: 'char' | 'varchar',
+  length: number,
   encoding?: Encoding,
 };
 
-export type BinaryDataType = {
-  baseType: 'blob' | 'binary' | 'varbinary',
-  length: number | null,
+export type StringWithoutLengthDataType = {
+  category: 'STRING',
+  baseType: 'text' | 'mediumtext' | 'longtext',
+  encoding?: Encoding,
 };
 
 export type EnumDataType = {
+  category: 'STRING',
   baseType: 'enum',
   values: Array<string>,
   encoding?: Encoding,
 };
 
-// These data types have no params
-export type OtherDataType = {
-  baseType: 'date' | 'year' | 'tinyblob' | 'mediumblob' | 'longblob' | 'json',
+export type StringDataType = StringWithLengthDataType | StringWithoutLengthDataType;
+
+export type StringOrEnumDataType = StringDataType | EnumDataType;
+
+export type BinaryWithLengthDataType = {
+  category: 'BINARY',
+  baseType: 'blob' | 'binary' | 'varbinary',
+  length: number | null,
+};
+
+export type BinaryWithoutLengthDataType = {
+  category: 'BINARY',
+  baseType: 'tinyblob' | 'mediumblob' | 'longblob',
+};
+
+export type BinaryDataType = BinaryWithLengthDataType | BinaryWithoutLengthDataType;
+
+export type JSONDataType = {
+  category: 'JSON',
+  baseType: 'json',
 };
 
 export type DataType =
-  | IntDataType
-  | RealDataType
-  | DateTimeDataType
-  | TextDataType
-  | BinaryDataType
+  | NumericDataType
+  | TemporalDataType
+  | StringDataType
   | EnumDataType
-  | OtherDataType;
+  | BinaryDataType
+  | JSONDataType;
 
 const DEFAULT_INT_LENGTHS = {
   bigint: 20,
@@ -67,7 +116,7 @@ const DEFAULT_INT_LENGTHS = {
   tinyint: 4,
 };
 
-function asInt(baseType: $PropertyType<IntDataType, 'baseType'>, params: string, options: string): IntDataType {
+function asInt(baseType: $PropertyType<IntegerDataType, 'baseType'>, params: string, options: string): IntegerDataType {
   const unsigned = /unsigned/i.test(options);
   const zeroFill = /zerofill/i.test(options);
 
@@ -80,10 +129,14 @@ function asInt(baseType: $PropertyType<IntDataType, 'baseType'>, params: string,
     if (unsigned) length -= 1;
   }
 
-  return { baseType, length, unsigned, zeroFill };
+  return { category: 'NUMERIC', subCategory: 'INTEGER', baseType, length, unsigned, zeroFill };
 }
 
-function asReal(baseType: $PropertyType<RealDataType, 'baseType'>, params: string, options: string): RealDataType {
+function asFixedPoint(
+  baseType: $PropertyType<FixedPointDataType, 'baseType'>,
+  params: string,
+  options: string,
+): FixedPointDataType {
   const unsigned = /unsigned/i.test(options);
   const zeroFill = /zerofill/i.test(options);
 
@@ -95,7 +148,26 @@ function asReal(baseType: $PropertyType<RealDataType, 'baseType'>, params: strin
     precision = { length, decimals };
   }
 
-  return { baseType, precision, unsigned, zeroFill };
+  return { category: 'NUMERIC', subCategory: 'FIXED_POINT', baseType, precision, unsigned, zeroFill };
+}
+
+function asFloatingPoint(
+  baseType: $PropertyType<FloatingPointDataType, 'baseType'>,
+  params: string,
+  options: string,
+): FloatingPointDataType {
+  const unsigned = /unsigned/i.test(options);
+  const zeroFill = /zerofill/i.test(options);
+
+  let precision = null;
+  if (params) {
+    const [first, second] = params.split(',').map((s) => s.trim());
+    const length = parseInt(first, 10);
+    const decimals = second ? parseInt(second, 10) : 0;
+    precision = { length, decimals };
+  }
+
+  return { category: 'NUMERIC', subCategory: 'FLOATING_POINT', baseType, precision, unsigned, zeroFill };
 }
 
 function asDateTime(
@@ -104,11 +176,24 @@ function asDateTime(
   // options: string,
 ): DateTimeDataType {
   const fsp = params ? parseInt(params, 10) : null;
-  return { baseType, fsp };
+  return { category: 'TEMPORAL', baseType, fsp };
 }
 
-function asText(baseType: $PropertyType<TextDataType, 'baseType'>, params: string, options: string): TextDataType {
-  let length = null;
+function asTimeOnly(baseType: $PropertyType<TimeOnlyDataType, 'baseType'>, params: string): TimeOnlyDataType {
+  const fsp = params ? parseInt(params, 10) : null;
+  return { category: 'TEMPORAL', baseType, fsp };
+}
+
+function asDateOnly(baseType: $PropertyType<DateOnlyDataType, 'baseType'>): DateOnlyDataType {
+  return { category: 'TEMPORAL', baseType };
+}
+
+function asStringWithLength(
+  baseType: $PropertyType<StringWithLengthDataType, 'baseType'>,
+  params: string,
+  options: string,
+): StringWithLengthDataType {
+  let length;
   if (params) {
     length = parseInt(params, 10);
   }
@@ -119,16 +204,27 @@ function asText(baseType: $PropertyType<TextDataType, 'baseType'>, params: strin
     length = 1;
   }
 
-  if (baseType === 'varchar' && !length) {
+  if (!length) {
     // VARCHAR without a length is invalid MySQL
     throw new Error('VARCHAR must have valid length, please use VARCHAR(n)');
   }
 
   const encoding = parseEncodingOptions(options);
-  return { baseType, length, encoding };
+  return { category: 'STRING', baseType, length, encoding };
 }
 
-function asBinary(baseType: $PropertyType<BinaryDataType, 'baseType'>, params: string): BinaryDataType {
+function asStringWithoutLength(
+  baseType: $PropertyType<StringWithoutLengthDataType, 'baseType'>,
+  options: string,
+): StringWithoutLengthDataType {
+  const encoding = parseEncodingOptions(options);
+  return { category: 'STRING', baseType, encoding };
+}
+
+function asBinary(
+  baseType: $PropertyType<BinaryWithLengthDataType, 'baseType'>,
+  params: string,
+): BinaryWithLengthDataType {
   let length = null;
   if (params) {
     length = parseInt(params, 10);
@@ -140,7 +236,13 @@ function asBinary(baseType: $PropertyType<BinaryDataType, 'baseType'>, params: s
     throw new Error('VARBINARY must have valid length, please use VARBINARY(n)');
   }
 
-  return { baseType, length };
+  return { category: 'BINARY', baseType, length };
+}
+
+function asBinaryWithoutLength(
+  baseType: $PropertyType<BinaryWithoutLengthDataType, 'baseType'>,
+): BinaryWithoutLengthDataType {
+  return { category: 'BINARY', baseType };
 }
 
 // TODO: Honestly, why are we not just doing this at the parser level?
@@ -172,7 +274,23 @@ function asEnum(baseType: $PropertyType<EnumDataType, 'baseType'>, params: strin
 
   const values = parseEnumValues(params);
   const encoding = parseEncodingOptions(options);
-  return { baseType, values, encoding };
+  return { category: 'STRING', baseType, values, encoding };
+}
+
+function asJSON(baseType: $PropertyType<JSONDataType, 'baseType'>): JSONDataType {
+  return { category: 'JSON', baseType: 'json' };
+}
+
+export function setEncoding<T: StringOrEnumDataType>(dataType: T, encoding: Encoding | void): T {
+  if (dataType.baseType === 'char' || dataType.baseType === 'varchar') {
+    return { ...dataType, encoding };
+  } else if (dataType.baseType === 'text' || dataType.baseType === 'mediumtext' || dataType.baseType === 'longtext') {
+    return { ...dataType, encoding };
+  } else if (dataType.baseType === 'enum') {
+    return { ...dataType, encoding };
+  }
+
+  throw new Error('Unknown string column: ' + dataType.baseType);
 }
 
 /**
@@ -207,21 +325,29 @@ export function parseDataType(type: string): DataType {
 
     case 'float':
     case 'double':
-    case 'decimal':
-      return asReal(baseType, params, options);
+      return asFloatingPoint(baseType, params, options);
 
-    // case 'date': // NOTE: "date" does not belong here! It's a "paramless" type.
-    case 'time':
+    case 'decimal':
+      return asFixedPoint(baseType, params, options);
+
     case 'timestamp':
     case 'datetime':
       return asDateTime(baseType, params /* , options */);
 
+    case 'time':
+      return asTimeOnly(baseType, params /* , options */);
+
+    case 'date':
+      return asDateOnly(baseType /* , params, options */);
+
     case 'char':
     case 'varchar':
+      return asStringWithLength(baseType, params, options);
+
     case 'text':
     case 'mediumtext':
     case 'longtext':
-      return asText(baseType, params, options);
+      return asStringWithoutLength(baseType, /* params, */ options);
 
     case 'binary':
     case 'varbinary':
@@ -231,13 +357,13 @@ export function parseDataType(type: string): DataType {
     case 'enum':
       return asEnum(baseType, params, options);
 
-    case 'date':
-    case 'year':
+    case 'json':
+      return asJSON(baseType /* params, options */);
+
     case 'tinyblob':
     case 'mediumblob':
     case 'longblob':
-    case 'json':
-      return { baseType };
+      return asBinaryWithoutLength(baseType);
 
     default:
       throw new Error(`Unrecognized MySQL data type: ${baseType}`);
@@ -280,11 +406,30 @@ export function formatDataType(dataType: DataType, tableEncoding: Encoding, full
       break;
 
     case 'char':
-    case 'varchar':
+    case 'varchar': {
+      params = dataType.length || '';
+
+      const encoding = dataType.encoding ?? tableEncoding;
+
+      // NOTE: This is some weird MySQL quirk... if an encoding is set
+      // explicitly, then the *collate* defines what gets displayed, otherwise
+      // the *charset* difference will determine it
+      let outputCharset = dataType.encoding !== undefined && dataType.encoding.collate !== tableEncoding.collate;
+      let outputCollation = encoding.collate !== getDefaultCollationForCharset(encoding.charset);
+
+      options = [
+        fullyResolved || outputCharset ? `CHARACTER SET ${encoding.charset}` : null,
+        fullyResolved || outputCollation ? `COLLATE ${encoding.collate}` : null,
+      ]
+        .filter(Boolean)
+        .join(' ');
+      break;
+    }
+
     case 'text':
     case 'mediumtext':
     case 'longtext': {
-      params = dataType.length || '';
+      params = '';
 
       const encoding = dataType.encoding ?? tableEncoding;